Code

+++ galaxy_zoo.py
@@ -33,7 +33,7 @@ for i, galaxy in enumerate(galaxies):
 
 
     # open the image and append it to the main list
-    image = mpimg.imread("/cosma7/data/durham/dc-howi1/project/project/Galaxy Zoo Images/gz2_images_all/" + galaxy)
+    image = mpimg.imread("/cosma7/data/durham/dc-howi1/project/project/Galaxy Zoo Images/gz2_images_all/" + galaxy).astype('float64')
 
     # normalise each band independently between 0 and 1
     image = normalise_independently(image)
